move datatype conversions to new module
